requirements.txt
fastapi==0.135.3
psycopg2-binary==2.9.11
pydantic==2.12.5
uvicorn==0.29.0
